Split view tests into 1 file per view
- Influenced by RSpec's view specs: https://relishapp.com/rspec/rspec-rails/v/2-3/docs/view-specs/view-spec
- Easier to navigate
